requirements.txt
#requirements.txt

fastapi==0.115.4
uvicorn==0.22.0
pydantic==2.9.2
groq==0.4.2
langchain==0.3.6
langchain-groq==0.2.1
langchain-community==0.3.5
langchain-huggingface==0.1.2
faiss-cpu==1.8.0.post1
python-dotenv==1.0.0
asyncio==3.4.3
pandas==2.1.4